bot2/requirements.txt
aiogram
torchvision
torch
Pillow
translate
matplotlib
numpy